Add utility function for merging class names
This commit adds a new utility function `cn` to the `utils.js` file in the `client/src/lib` directory. The `cn` function uses the `clsx` and `twMerge` functions to merge multiple class names into a single string. This utility function will be helpful for managing class names in the project.